requirements.txt
Django
djangorestframework
djangorestframework-jwt
pytest
pytest-django
pytest-cov
